+++ src/components/Auth.vue
